--- src/components/inputs/rich-text-input/plugins/index.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export { RenderBlockPlugin, RenderMarkPlugin } from './plugins';
