+++ src/components/inputs/rich-text-input/rich-text-input.story.js
@@ -7,9 +7,11 @@ import Spacings from '../../spacings';
 import Section from '../../../../.storybook/decorators/section';
 import RichTextInput from './rich-text-input';
 import Readme from './README.md';
-import jsonValue from './value';
+import jsonValue from './testing/json-values/large-value';
 import TextInput from '../text-input';
+
 // Create our initial value...
+
 const initialValue = Value.fromJSON(jsonValue);
 storiesOf('Components|Inputs', module)
   .addDecorator(withKnobs)
