+++ src/components/inputs/rich-text-input/testing/json-values/minimal-value.json
