--- src/components/inputs/rich-text-input/utils/index.js
@@ -1,1 +1,0 @@
-export { default } from './rules';
