--- src/components/internals/rich-text-body/utils/index.js
@@ -1,1 +1,0 @@
-export { default } from './rules';
