+++ src/components/internals/rich-text-plugins/plugins/list.js
@@ -1,16 +1,14 @@
 import React from 'react';
+import { BLOCK_TAGS } from '../../rich-text-utils/tags';
 
-const NUMBERED_LIST = 'numbered-list';
-const BULLETED_LIST = 'bulleted-list';
-const LIST_ITEM = 'list-item';
-const DEFAULT_NODE = 'paragraph';
+const DEFAULT_NODE = BLOCK_TAGS.p;
 
 const hasBlock = (type, editor) =>
   editor.value.blocks.some(node => node.type === type);
 
 const toggle = (editor, typeName) => {
   // Handle the extra wrapping required for list buttons.
-  const isList = hasBlock('list-item', editor);
+  const isList = hasBlock(BLOCK_TAGS.li, editor);
   const isType = editor.value.blocks.some(block => {
     return Boolean(
       editor.value.document.getClosest(
@@ -23,14 +21,14 @@ const toggle = (editor, typeName) => {
   if (isList && isType) {
     editor
       .setBlocks(DEFAULT_NODE)
-      .unwrapBlock(BULLETED_LIST)
-      .unwrapBlock(NUMBERED_LIST);
+      .unwrapBlock(BLOCK_TAGS.ul)
+      .unwrapBlock(BLOCK_TAGS.ol);
   } else if (isList) {
     editor
-      .unwrapBlock(typeName === BULLETED_LIST ? NUMBERED_LIST : BULLETED_LIST)
+      .unwrapBlock(typeName === BLOCK_TAGS.ul ? BLOCK_TAGS.ol : BLOCK_TAGS.ul)
       .wrapBlock(typeName);
   } else {
-    editor.setBlocks(LIST_ITEM).wrapBlock(typeName);
+    editor.setBlocks(BLOCK_TAGS.li).wrapBlock(typeName);
   }
 };
 
@@ -42,7 +40,7 @@ const query = (editor, typeName) => {
       editor.value.blocks.first().key
     );
     isActive =
-      hasBlock(LIST_ITEM, editor) && parent && parent.type === typeName;
+      hasBlock(BLOCK_TAGS.li, editor) && parent && parent.type === typeName;
   }
   return isActive;
 };
@@ -54,11 +52,11 @@ const ListPlugin = () => {
         const { attributes, children, node } = props;
 
         switch (node.type) {
-          case BULLETED_LIST:
+          case BLOCK_TAGS.ul:
             return <ul {...attributes}>{children}</ul>;
-          case LIST_ITEM:
+          case BLOCK_TAGS.li:
             return <li {...attributes}>{children}</li>;
-          case NUMBERED_LIST:
+          case BLOCK_TAGS.ol:
             return <ol {...attributes}>{children}</ol>;
           default:
             return next();
@@ -69,18 +67,18 @@ const ListPlugin = () => {
           if (!editor.value.selection.isFocused) {
             editor.focus();
           }
-          toggle(editor, BULLETED_LIST);
+          toggle(editor, BLOCK_TAGS.ul);
         },
         toggleNumberedListBlock: editor => {
           if (!editor.value.selection.isFocused) {
             editor.focus();
           }
-          toggle(editor, NUMBERED_LIST);
+          toggle(editor, BLOCK_TAGS.ol);
         },
       },
       queries: {
-        hasBulletedListBlock: editor => query(editor, BULLETED_LIST),
-        hasNumberedListBlock: editor => query(editor, NUMBERED_LIST),
+        hasBulletedListBlock: editor => query(editor, BLOCK_TAGS.ul),
+        hasNumberedListBlock: editor => query(editor, BLOCK_TAGS.ol),
       },
     },
   ];
