+++ src/components/internals/rich-text-plugins/rich-text-plugins.js
@@ -5,7 +5,7 @@ import ListPlugin from './plugins/list';
 import MarkPlugin from './plugins/mark';
 import { RenderMarkPlugin, RenderBlockPlugin } from './plugins';
 import PlaceholderPlugin from './plugins/placeholder';
-import { BLOCK_TAGS, MARK_TAGS } from '../rich-text-body/utils/rules';
+import { BLOCK_TAGS, MARK_TAGS } from '../rich-text-utils/tags';
 
 const isEmpty = value => value.document.text === '';
