--- src/components/internals/rich-text-utils/index.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export { isEmpty } from './rich-text-utils';
