+++ src/components/internals/rich-text-utils/is-empty.js
@@ -1,0 +1,3 @@
+const isEmpty = value => value.document.text === '';
+
+export default isEmpty;
