--- src/components/internals/rich-text-utils/rich-text-utils.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export const isEmpty = value => value.document.text === '';
