+++ src/components/internals/rich-text-utils/tags.js
@@ -1,0 +1,23 @@
+export const BLOCK_TAGS = {
+  blockquote: 'block-quote',
+  p: 'paragraph',
+  h1: 'heading-one',
+  h2: 'heading-two',
+  h3: 'heading-three',
+  h4: 'heading-four',
+  h5: 'heading-five',
+  pre: 'code',
+  li: 'list-item',
+  ol: 'numbered-list',
+  ul: 'bulleted-list',
+};
+
+// Add a dictionary of mark tags.
+export const MARK_TAGS = {
+  em: 'italic',
+  strong: 'bold',
+  u: 'underlined',
+  sup: 'superscript',
+  sub: 'subscript',
+  del: 'strikethrough',
+};
